+++ packages/launch-kit/src/index.ts
@@ -3,4 +3,4 @@ export * from "./utility-manifest.js";
 export * from "./readiness.js";
 export * from "./validate.js";
 export * from "./bundle.js";
-export * from "./adapters/based-bid.js";
+export * from "./adapters/launchpad.js";